(Update) AWS migration complete

+++ wp-content/plugins/ultimate-member/includes/core/class-permalinks.php
@@ -113,37 +113,43 @@ if ( ! class_exists( 'um\core\Permalinks' ) ) {
 				 isset( $_REQUEST['user_id'] ) && is_numeric( $_REQUEST['user_id'] ) ) { // valid token
 
 				$user_id = absint( $_REQUEST['user_id'] );
+				if ( is_user_logged_in() && get_current_user_id() !== $user_id ) {
+					// Cannot activate another user account. Please log out and try again.
+					wp_safe_redirect( um_user_profile_url( get_current_user_id() ) );
+					exit;
+				}
+
 				delete_option( "um_cache_userdata_{$user_id}" );
 
 				$account_secret_hash = get_user_meta( $user_id, 'account_secret_hash', true );
 				if ( empty( $account_secret_hash ) || strtolower( sanitize_text_field( $_REQUEST['hash'] ) ) !== strtolower( $account_secret_hash ) ) {
-					wp_die( __( 'This activation link is expired or have already been used.', 'ultimate-member' ) );
+					wp_safe_redirect( add_query_arg( 'err', 'activation_link_used', um_get_core_page( 'login' ) ) );
+					exit;
 				}
 
 				$account_secret_hash_expiry = get_user_meta( $user_id, 'account_secret_hash_expiry', true );
 				if ( ! empty( $account_secret_hash_expiry ) && time() > $account_secret_hash_expiry ) {
-					wp_die( __( 'This activation link is expired.', 'ultimate-member' ) );
+					wp_safe_redirect( add_query_arg( 'err', 'activation_link_expired', um_get_core_page( 'login' ) ) );
+					exit;
 				}
 
+				$redirect              = um_get_core_page( 'login', 'account_active' );
+				$set_password_required = get_user_meta( $user_id, 'um_set_password_required', true );
+
 				um_fetch_user( $user_id );
-				UM()->user()->approve();
+				UM()->common()->users()->approve( $user_id, true );
+				if ( ! empty( $set_password_required ) ) {
+					$redirect = um_user( 'password_reset_link' );
+				}
 				um_reset_user();
 
-				$user_role = UM()->roles()->get_priority_user_role( $user_id );
+				$user_role      = UM()->roles()->get_priority_user_role( $user_id );
 				$user_role_data = UM()->roles()->role_data( $user_role );
 
 				// log in automatically
 				$login = ! empty( $user_role_data['login_email_activate'] ); // Role setting "Login user after validating the activation link?"
 				if ( ! is_user_logged_in() && $login ) {
-					$user = get_userdata( $user_id );
-
-					// update wp user
-					wp_set_current_user( $user_id, $user->user_login );
-					wp_set_auth_cookie( $user_id );
-
-					ob_start();
-					do_action( 'wp_login', $user->user_login, $user );
-					ob_end_clean();
+					UM()->user()->auto_login( $user_id );
 				}
 
 				/**
@@ -167,7 +173,9 @@ if ( ! class_exists( 'um\core\Permalinks' ) ) {
 				 */
 				do_action( 'um_after_email_confirmation', $user_id );
 
-				$redirect = empty( $user_role_data['url_email_activate'] ) ? um_get_core_page( 'login', 'account_active' ) : trim( $user_role_data['url_email_activate'] ); // Role setting "URL redirect after e-mail activation"
+				if ( empty( $set_password_required ) ) {
+					$redirect = empty( $user_role_data['url_email_activate'] ) ? um_get_core_page( 'login', 'account_active' ) : trim( $user_role_data['url_email_activate'] ); // Role setting "URL redirect after email activation"
+				}
 				$redirect = apply_filters( 'um_after_email_confirmation_redirect', $redirect, $user_id, $login );
 
 				exit( wp_redirect( $redirect ) );
@@ -488,86 +496,6 @@ if ( ! class_exists( 'um\core\Permalinks' ) ) {
 			}
 
 			return $user_in_url;
-		}
-
-		/**
-		 * Get action url for admin use
-		 *
-		 * @param $action
-		 * @param $subaction
-		 *
-		 * @deprecated 2.6.9
-		 *
-		 * @return mixed|string|void
-		 */
-		public function admin_act_url( $action, $subaction ) {
-			_deprecated_function( __METHOD__, '2.6.9' );
-			$url = add_query_arg(
-				array(
-					'um_adm_action' => $action,
-					'sub'           => $subaction,
-					'user_id'       => um_user( 'ID' ),
-					'_wpnonce'      => wp_create_nonce( $action ),
-				)
-			);
-			return $url;
-		}
-
-		/**
-		 * SEO canonical href bugfix
-		 *
-		 * @deprecated since version 2.1.7
-		 *
-		 * @todo remove since 2.7.0
-		 * @see function um_profile_remove_wpseo()
-		 */
-		public function um_rel_canonical_() {
-			_deprecated_function( __METHOD__, '2.1.7', 'um_profile_remove_wpseo()' );
-			global $wp_the_query;
-
-			if ( ! is_singular() )
-				return;
-
-			/**
-			 * UM hook
-			 *
-			 * @type filter
-			 * @title um_allow_canonical__filter
-			 * @description Allow canonical
-			 * @input_vars
-			 * [{"var":"$allow_canonical","type":"bool","desc":"Allow?"}]
-			 * @change_log
-			 * ["Since: 2.0"]
-			 * @usage
-			 * <?php add_filter( 'um_allow_canonical__filter', 'function_name', 10, 1 ); ?>
-			 * @example
-			 * <?php
-			 * add_filter( 'um_allow_canonical__filter', 'my_allow_canonical', 10, 1 );
-			 * function my_allow_canonical( $allow_canonical ) {
-			 *     // your code here
-			 *     return $allow_canonical;
-			 * }
-			 * ?>
-			 */
-			$enable_canonical = apply_filters( "um_allow_canonical__filter", true );
-
-			if( ! $enable_canonical )
-				return;
-
-			if ( !$id = $wp_the_query->get_queried_object_id() )
-				return;
-
-			if ( UM()->config()->permalinks['user'] == $id ) {
-				$link = esc_url( $this->get_current_url() );
-				echo "<link rel='canonical' href='$link' />\n";
-				return;
-			}
-
-			$link = get_permalink( $id );
-			if ( $page = get_query_var( 'cpage' ) ) {
-				$link = get_comments_pagenum_link( $page );
-				echo "<link rel='canonical' href='$link' />\n";
-			}
 		}
 	}
 }
